More renames, cleanup after the previous PR.
