+++ build_runner/test/common/common.dart
@@ -10,11 +10,11 @@ import 'package:crypto/crypto.dart';
 export 'package:build_test/build_test.dart';
 export 'package:build_test/src/internal_test_reader_writer.dart';
 
+export 'build_packages.dart';
 export 'build_runner_tester.dart';
 export 'builders.dart';
 export 'fixture_packages.dart';
 export 'matchers.dart';
-export 'package_graphs.dart';
 export 'test_phases.dart';
 
 Digest computeDigest(AssetId id, String contents) {
